+++ Calculantique/src/babylonien/Clou.java
@@ -1,7 +1,7 @@
 package babylonien;
 
 public class Clou {
-	private int valeur;
+	private long valeur;
 	
 	public Clou(int saVal){
 		this.valeur=saVal;
@@ -11,13 +11,13 @@ public class Clou {
 		this(1);
 	}
 
-	public int getValeur(){
+	public long getValeur(){
 		return this.valeur;
 	}
 	
-	public String nbClou1 (int nombre){
+	public String nbClou1 (long nombre){
 		
-		int i = nombre%10;
+		long i = nombre%10;
 		String nombre2 = "";
 		
 		while (i > 0)
